:truck: Move types up to parent
