+++ apricot/oauth/__init__.py
@@ -1,7 +1,8 @@
+from apricot.types import LDAPAttributeDict, LDAPControlTuple
+
 from .enums import OAuthBackend
 from .microsoft_entra_client import MicrosoftEntraClient
 from .oauth_client import OAuthClient
-from .types import LDAPAttributeDict, LDAPControlTuple
 
 OAuthClientMap = {OAuthBackend.MICROSOFT_ENTRA: MicrosoftEntraClient}
